+++ mjml/helpers/width_parser.py
@@ -29,17 +29,10 @@ def widthParser(width, parseFloatToInt=True):
     width_str = str(width)
     match = unitRegex.search(width_str)
     widthUnit = match.group(1)
+
     if (widthUnit == '%') and not parseFloatToInt:
-        parser = float
+        parsed_width = strip_unit(width_str)
     else:
-        parser = int
-    width = strip_unit(width_str)
-    parsed_width = parser(width)
-    # LATER: somehow JS works differently here (as it does not have a strict
-    # type sytem). parseFloat() might return a number without fractional but
-    # python does.
-    width_int = int(width)
-    if parsed_width == width_int:
-        parsed_width = width_int
+        parsed_width = int(strip_unit(width_str))
 
     return WidthUnit(width=parsed_width, unit=widthUnit)